Decode audio as float32 to reduce peak memory usage
Halves the size of the raw sample buffer and all derived librosa
arrays (chroma/CQT/tempogram etc.) for long tracks, with no change
to output values. Aimed at reducing OOM risk on Railway's
memory-constrained instances for multi-minute files.

+++ analyze.py
@@ -193,7 +193,7 @@ def compute_stereo_metrics(data):
 
 
 def analyze(audio_bytes):
-    data, sr = sf.read(io.BytesIO(audio_bytes), always_2d=True, dtype='float64')
+    data, sr = sf.read(io.BytesIO(audio_bytes), always_2d=True, dtype='float32')
     if data.shape[1] == 1:
         data = np.repeat(data, 2, axis=1)
     y = librosa.to_mono(data.T)
